requirement.txt
Django==5.0.7
djangorestframework==3.16.0
pandas==2.3.1
gunicorn==23.0.0
django-cors-headers==4.7.0
openpyxl==3.1.2